complete paths with any substring, anywhere
Type <tmux prefix>* and get a list of paths that will be inserted if you
select one.

NOTE: To reduce the number of paths, it only lists paths that are not
ignored by git, so currently it is meant to be used in repos.

TODO: Right now if too many paths are generated, it won't show any
pop-up. We should make it so it just cuts off the number of paths so as
many as possible get displayed.