The text on which to do the selection can come from copy mode
The only requirement is that the cursor be on the top line, that way we
can parse the history line number from the cursor line (I didn't find a
better way to do this, maybe this number is obtainable directly
somehow). So the cursor needs to be forced to the top if in copy mode
before doing the capture.